refactor: DDD 적용, 코드 리팩토링
