+++ src/main/java/com/yunbok/searchapi/v1/authentication/vo/request/AccessTokenRequest.java
@@ -1,0 +1,6 @@
+package com.yunbok.searchapi.v1.authentication.vo.request;
+
+public record AccessTokenRequest (
+        String account
+) {}
+
