+++ src/main/java/com/yunbok/searchapi/v1/common/dto/ErrorResponse.java
@@ -1,22 +1,12 @@
 package com.yunbok.searchapi.v1.common.dto;
 
 import com.yunbok.searchapi.v1.common.define.ResponseCode;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class ErrorResponse implements Response {
-
-    private int code;
-    private String message;
-
-    private ErrorResponse(int code, String message) {
-        this.code = code;
-        this.message = message;
-    }
-
+public record ErrorResponse (
+        int code,
+        String message
+)
+{
     public static ErrorResponse responseOf(ResponseCode responseCode) {
         return new ErrorResponse(responseCode.getCode(), responseCode.getMessage());
     }
