--- src/main/java/com/yunbok/searchapi/v1/search/dto/BookResponse.java
@@ -1,12 +1,0 @@
-package com.yunbok.searchapi.v1.search.dto;
-
-public record BookResponse(
-        String library,
-        String location,
-        String registrationNumber,
-        String title,
-        String author,
-        String publisher,
-        String publicationYear,
-        String callNumber
-) {}
