--- src/main/java/com/yunbok/searchapi/v1/search/dto/SearchRequest.java
@@ -1,9 +1,0 @@
-package com.yunbok.searchapi.v1.search.dto;
-
-public record SearchRequest(
-        String query,
-        String sort,
-        int page,
-        int pageSize
-) {}
-
