+++ src/main/java/com/yunbok/searchapi/v1/search/vo/SearchRequest.java
@@ -1,0 +1,9 @@
+package com.yunbok.searchapi.v1.search.vo;
+
+public record SearchRequest(
+        String query,
+        String sort,
+        int page,
+        int pageSize
+) {}
+
