+++ src/test/java/com/yunbok/searchapi/v1/authentication/controller/AuthenticationControllerTest.java
@@ -1,13 +1,12 @@
 package com.yunbok.searchapi.v1.authentication.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.yunbok.searchapi.v1.authentication.dto.request.AccessTokenRequest;
-import com.yunbok.searchapi.v1.authentication.dto.request.ApiKeyRequest;
-import com.yunbok.searchapi.v1.authentication.dto.response.AccessTokenResponse;
-import com.yunbok.searchapi.v1.authentication.dto.response.ApiKeyResponse;
-import com.yunbok.searchapi.v1.authentication.exception.ApiAuthenticationException;
-import com.yunbok.searchapi.v1.authentication.service.AuthenticationService;
-import com.yunbok.searchapi.v1.common.define.ResponseCode;
+import com.yunbok.searchapi.v1.authentication.vo.response.AccessTokenResponse;
+import com.yunbok.searchapi.v1.authentication.presentation.AuthenticationController;
+import com.yunbok.searchapi.v1.authentication.vo.request.AccessTokenRequest;
+import com.yunbok.searchapi.v1.authentication.vo.request.ApiKeyRequest;
+import com.yunbok.searchapi.v1.authentication.vo.response.ApiKeyResponse;
+import com.yunbok.searchapi.v1.authentication.application.AuthenticationService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -18,16 +17,12 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.RequestBuilder;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.refEq;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.refEq;
-import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(AuthenticationController.class)
@@ -50,7 +45,7 @@ public class AuthenticationControllerTest {
         String password = "testPassword";
         String apiKey = "testApiKey";
 
-        ApiKeyRequest apiKeyRequest = ApiKeyRequest.requestOf(account, password);
+        ApiKeyRequest apiKeyRequest = new ApiKeyRequest(account, password);
         ApiKeyResponse apiKeyResponse = ApiKeyResponse.successOf(apiKey);
 
         when(authenticationService.getApiKey(refEq(apiKeyRequest))).thenReturn(apiKeyResponse);
@@ -66,14 +61,16 @@ public class AuthenticationControllerTest {
         String responseContent = mvcResult.getResponse().getContentAsString();
         ApiKeyResponse responseDto = new ObjectMapper().readValue(responseContent, ApiKeyResponse.class);
 
-        assertEquals(apiKeyResponse.getApiKey(), responseDto.getApiKey());
+        assertEquals(apiKeyResponse.apikey(), responseDto.apikey());
     }
 
     @Test
     public void testGetAccessToken() throws Exception {
         // given
-        AccessTokenRequest request = AccessTokenRequest.requestOf("testAccount");
-        AccessTokenResponse response = AccessTokenResponse.responseOf("accessToken", 100000L);
+        AccessTokenRequest request = new AccessTokenRequest("testAccount");
+        AccessTokenResponse response = AccessTokenResponse.generateTokenOf(
+                "accessToken",
+                100000L);
 
         when(authenticationService.getAccessToken(any(), any())).thenReturn(response);
