+++ src/test/java/com/yunbok/searchapi/v1/authentication/util/ApiKeyGeneratorTest.java
@@ -1,6 +1,5 @@
 package com.yunbok.searchapi.v1.authentication.util;
 
-import com.yunbok.searchapi.v1.authentication.service.ApiKeyService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
